LRNG - add random.c entropy source support
The random.c implementation can be used as an entropy source by the
LRNG. This support can be enabled at compile time.

The entropy rate can be set at compile time which is only applied:

- once the random.c considers itself fully seeded, and

- the kernel does not operate in FIPS mode (i.e. fips=1 is not set at
  the kernel command line)

If one of these properties is not set, the ES will obtain data, but will
credit it with zero bits of entropy. For the first bullet, it is clear
why it will have zero bits of entropy. But for the second property, this
is set because the random.c is not operating SP800-90B compliant and
thus must be treated to not deliver any entropy in FIPS mode.

Change-Id: I44a66b17dc9cbc62c1c9a457ee444b48c1546f1a
Signed-off-by: Stephan Mueller <smueller@chronox.de>
Signed-off-by: farizmaul <nickshade88@gmail.com>